breaking: ha: client: switch to device based discovery
Currently we send discovery messages per-entity.
As we send all entities for a device at once anyways, it makes sense to
switch to per-device, to optimize data flow and make migration of entity
changes way easier.

BREAKING: This is a breaking change: We keep all unique_ids as-is, this
has some effects:
- Devices in HA are not duplicated
- History from previous discovered devices is preserved
- Users have to delete previous MQTT devices first to make discovery
work
Latter one results in the change being breaking.